+++ app/src/main/java/com/soundcloud/android/model/TrackProperty.java
@@ -1,5 +1,5 @@
 package com.soundcloud.android.model;
 
 public final class TrackProperty {
-    public static final Property<Long> PLAY_COUNT = Property.of(Long.class);
+    public static final Property<Integer> PLAY_COUNT = Property.of(Integer.class);
 }
